Relocate the L-J models
Makes a tiny improvement to compilation speed
